v2.5.0 hotfix (before jp translation)
- fixes translation error for custom items
- all translations fixed and up to date
- fixes audio volumes in one of the archives (now confirmed working)
- fixed that bug on the eye finally, which means a certain friendly face will not be waiting for you any longer.
- fixes some things related to solanum
- fixes a handful of null reference exceptions
- a certain vision torch can now be viewed from the bottom of a stairwell
- fixes a log condition not working properly for a certain deserted traveler

+++ TheStrangerTheyAre/GhostBirdHandlerTSTA.cs
@@ -10,19 +10,32 @@ namespace TheStrangerTheyAre
         private GameObject[] ghostBirds; // array of all ghostbirds, set in unity
         [SerializeField]
         private GameObject[] initialPos; // array for storing each starting position
-        private bool areDeadZone1 = false; // boolean that tests if the ghosts have died.
-        private bool areDeadZone2 = false; // boolean that tests if the ghosts have died.
-        private bool isCaught = false; // boolean to check if player is caught
+        private bool areDeadZone1; // boolean that tests if the ghosts have died.
+        private bool areDeadZone2; // boolean that tests if the ghosts have died.
+        private bool isCaught; // boolean to check if player is caught
 
         void Awake()
         {
-            for (int i = 0; i < ghostBirds.Length; i++)
-            {
-                initialPos[i].transform.position = ghostBirds[i].transform.position; // gets starting pos of each ghostbird
-                initialPos[i].transform.rotation = ghostBirds[i].transform.rotation; // gets starting rotation of each ghostbird
-            };
             GlobalMessenger.AddListener("ExitDreamWorld", OnExitDreamWorld); // checks if player leaves the sim
             GlobalMessenger.AddListener("PlayerGrabbedByGhost", OnPlayerGrabbedByGhost); // checks if player is caught by ghostbird
+        }
+
+        void Start()
+        {
+            // set all bools to false by default
+            areDeadZone1 = false;
+            areDeadZone2 = false;
+            isCaught = false;
+
+            // get ghost positions
+            for (int i = 0; i < ghostBirds.Length; i++)
+            {
+                if (initialPos[i] != null)
+                {
+                    initialPos[i].transform.position = ghostBirds[i].transform.position; // gets starting pos of each ghostbird
+                    initialPos[i].transform.rotation = ghostBirds[i].transform.rotation; // gets starting rotation of each ghostbird
+                }
+            }
         }
 
         void OnExitDreamWorld()
@@ -30,10 +43,13 @@ namespace TheStrangerTheyAre
             for (int i = 0; i < ghostBirds.Length; i++)
             {
                 isCaught = false;
-                ghostBirds[i].transform.position = initialPos[i].transform.position; // on dw exit, sets to starting pos of each ghostbird
-                ghostBirds[i].transform.rotation = initialPos[i].transform.rotation; // on dw exit, sets to starting rotation of each ghostbird
-                ghostBirds[i].GetComponent <GhostController>().SetLanternConcealed(true, true);
-                ghostBirds[i].GetComponentInChildren<Animator>().Play("Ghostbird_Idle_Unaware", 0); // on dw exit, sets each ghostbird animation to idle
+                if (ghostBirds[i] != null)
+                {
+                    ghostBirds[i].transform.position = initialPos[i].transform.position; // on dw exit, sets to starting pos of each ghostbird
+                    ghostBirds[i].transform.rotation = initialPos[i].transform.rotation; // on dw exit, sets to starting rotation of each ghostbird
+                    ghostBirds[i].GetComponent<GhostController>().SetLanternConcealed(true, true);
+                    ghostBirds[i].GetComponentInChildren<Animator>().Play("Ghostbird_Idle_Unaware", 0); // on dw exit, sets each ghostbird animation to idle
+                }
             }
         }
 
@@ -44,8 +60,6 @@ namespace TheStrangerTheyAre
 
         void Update()
         {
-
-            
             if (Locator._toolModeSwapper.GetItemCarryTool().GetHeldItem() is DreamLanternItem lantern && lantern._lanternController._lit)
             {
                 if (OWInput.IsPressed(InputLibrary.toolActionPrimary, InputMode.Character) && OWInput.IsPressed(InputLibrary.toolActionSecondary, InputMode.Character))
@@ -62,15 +76,21 @@ namespace TheStrangerTheyAre
             {
                 for (int i = 0; i < 2; i++)
                 {
-                    ghostBirds[i].GetComponent<GhostBrain>().Die(); // kill ghostbirds linked to zone 1
-                    ghostBirds[i].SetActive(false);
+                    if (ghostBirds[i] != null)
+                    {
+                        ghostBirds[i].GetComponent<GhostBrain>().Die(); // kill ghostbirds linked to zone 1
+                        ghostBirds[i].SetActive(false);
+                    }
                 };
                 areDeadZone1 = true; // set dead boolean for zone 1 true
             } else if (TimeLoop.GetSecondsElapsed() > 1230 && !areDeadZone2)
             {
                 for (int i = 2; i < 8; i++)
                 {
-                    ghostBirds[i].GetComponent<GhostBrain>().Die(); // kill ghostbirds linked to zone 2
+                    if (ghostBirds[i] != null)
+                    {
+                        ghostBirds[i].GetComponent<GhostBrain>().Die(); // kill ghostbirds linked to zone 2
+                    }
                 };
                 areDeadZone2 = true; // set dead boolean for zone 2 true
             }
@@ -79,8 +99,11 @@ namespace TheStrangerTheyAre
         {
             foreach (var bird in ghostBirds)
             {
-                Vector3 smallTrigger = new Vector3(1, 1, 1);
-                bird.transform.Find("ContactTrigger/ContactTrigger_Core").gameObject.transform.localScale = smallTrigger;
+                if (bird != null)
+                {
+                    Vector3 smallTrigger = new Vector3(1, 1, 1);
+                    bird.transform.Find("ContactTrigger/ContactTrigger_Core").gameObject.transform.localScale = smallTrigger;
+                }
             }
         }
 
@@ -88,15 +111,18 @@ namespace TheStrangerTheyAre
         {
             foreach (var bird in ghostBirds)
             {
-                if (bird.GetComponentInChildren<CompoundLightSensor>().IsIlluminated() && !isCaught)
+                if (bird != null)
                 {
-                    Vector3 giantTrigger = new Vector3(7, 7, 7);
-                    bird.transform.Find("ContactTrigger/ContactTrigger_Core").gameObject.transform.localScale = giantTrigger;
-                }
-                else
-                {
-                    Vector3 bigTrigger = new Vector3(2, 2, 2);
-                    bird.transform.Find("ContactTrigger/ContactTrigger_Core").gameObject.transform.localScale = bigTrigger;
+                    if (bird.GetComponentInChildren<CompoundLightSensor>().IsIlluminated() && !isCaught)
+                    {
+                        Vector3 giantTrigger = new Vector3(7, 7, 7);
+                        bird.transform.Find("ContactTrigger/ContactTrigger_Core").gameObject.transform.localScale = giantTrigger;
+                    }
+                    else
+                    {
+                        Vector3 bigTrigger = new Vector3(2, 2, 2);
+                        bird.transform.Find("ContactTrigger/ContactTrigger_Core").gameObject.transform.localScale = bigTrigger;
+                    }
                 }
             }
         }
